refactor(activities): move página de atividades para rota dinâmica e ajusta dados da sidebar
