+++ src/features/activities/components/upload/activity-upload-modal.tsx
@@ -70,7 +70,6 @@ export const ActivityUploadModal = ({ isOpen, onClose, onDataParsed}: ActivityUp
 
     setLoading(true);
 
-    type CsvRow = Record<string, string>;
 
     Papa.parse(file, {
       header: true,
