+++ src/utils/sidebar-data.tsx
@@ -2,7 +2,7 @@ import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { AuthUserWithProfileT } from "@/types/auth";
 import { ClassroomT } from "@/types/classrooms";
 import { LucideIcon } from "lucide-react";
-import createClassroomNavigation from "./classroom-navigation";
+// import createClassroomNavigation from "./classroom-navigation";
 
 const createClassroomNavigation = (
   classrooms: ClassroomT[],
